+++ docs/asset-manifest.json
@@ -1,7 +1,7 @@
 {
   "files": {
-    "main.css": "/php-wasm/static/css/main.0fa749dd.css",
-    "main.js": "/php-wasm/static/js/main.5f2b3be2.js",
+    "main.css": "/php-wasm/static/css/main.211eeb60.css",
+    "main.js": "/php-wasm/static/js/main.eb710935.js",
     "static/media/php-web.mjs.wasm": "/php-wasm/static/media/php-web.mjs.57264d13341339b05062.wasm",
     "worker-xquery.js": "/php-wasm/16470d85f9e6b184411b622ce1783fd0.js",
     "worker-javascript.js": "/php-wasm/70896c45cb42a1a16168cef789116031.js",
@@ -491,11 +491,11 @@
     "ftl.js": "/php-wasm/f76ee9c8abfdd96fb9d70116d40435d5.js",
     "d.js": "/php-wasm/c98e74fc97b04fe8bf43dcdff549afcf.js",
     "mode-text.js": "/php-wasm/0ffb18fb70c87335edee31a479f58a43.js",
-    "main.0fa749dd.css.map": "/php-wasm/static/css/main.0fa749dd.css.map",
-    "main.5f2b3be2.js.map": "/php-wasm/static/js/main.5f2b3be2.js.map"
+    "main.211eeb60.css.map": "/php-wasm/static/css/main.211eeb60.css.map",
+    "main.eb710935.js.map": "/php-wasm/static/js/main.eb710935.js.map"
   },
   "entrypoints": [
-    "static/css/main.0fa749dd.css",
-    "static/js/main.5f2b3be2.js"
+    "static/css/main.211eeb60.css",
+    "static/js/main.eb710935.js"
   ]
 }
